% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr.R
\name{covspr}
\alias{covspr}
\title{Calculate semi-partial correlations from a 3x3 correlation matrix}
\usage{
covspr(f, data)
}
\arguments{
\item{f}{Formula. Two variables to compute semi-partial correlation. LHS MUST be dep. variable.
RHS must be ind. variable of interest}

\item{data}{Correlation matrix}
}
\value{
Semi-partial correlation value
}
\description{
Calculate semi-partial correlations from a 3x3 correlation matrix
}
\examples{
cm <- matrix(data=c(1, 0.2, 0.3, 0.2, 1, -0.6, 0.3, -0.6, 1), nrow=3, ncol=3)
colnames(t1) <- c("y", "x1", "x2")
rownames(t1) <- c("y", "x1", "x2")
covspr(y ~ x1, data = cm)

}
