% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaR2.R
\name{deltaR2}
\alias{deltaR2}
\title{Calculate delta R squared for two linear models}
\usage{
deltaR2(fullm, restrim, data)
}
\arguments{
\item{fullm}{The full model formula}

\item{restrim}{The restricted model formula}

\item{data}{data frame of data}
}
\value{
Returns a numeric vector of F test
}
\description{
Calculate delta R squared for two linear models
}
