% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-significance_testing.R
\name{fmr}
\alias{fmr}
\title{Full model F-test on multiple regression model}
\usage{
fmr(f, data)
}
\arguments{
\item{f}{Formula for model}

\item{data}{Data.frame of data}
}
\value{
F-test
}
\description{
Full model F-test on multiple regression model
}
\examples{
d <- mrc::simdat(n = 100, r = 0.4)
fmr(y ~ x1 + x2, data = d)

}
