% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-significance_testing.R
\name{fwrite}
\alias{fwrite}
\title{Takes full model f-test and spits out a written version of the test.}
\usage{
fwrite(test)
}
\arguments{
\item{test}{Output from \code{fmr()}}
}
\value{
Prints out statement in R Console
}
\description{
Takes full model f-test and spits out a written version of the test.
}
