% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-significance_testing.R
\name{mms}
\alias{mms}
\title{Model mean squared}
\usage{
mms(f, data)
}
\arguments{
\item{f}{Formula for model}

\item{data}{Data.frame of data}
}
\value{
MS value
}
\description{
Model mean squared
}
\examples{
d <- mrc::simdat(n = 100, r = 0.4)
mms(y ~ x1 + x2, data = d)

}
