% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_extract.R
\name{mrm_extract}
\alias{mrm_extract}
\title{Extract parameters from a fitted model}
\usage{
mrm_extract(rc_fit)
}
\arguments{
\item{rc_fit}{A fitted model object.}
}
\value{
A list containing the center, lower, and upper bounds of the parameters.
}
\description{
This function extracts the parameters from a fitted model object.
}
\details{
The function extracts the parameters from the fitted model object and returns them in a list.
}
