% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_optimal_mix.R
\name{mrm_optimal_mix}
\alias{mrm_optimal_mix}
\title{Optimal Mix Function
This function computes the optimal mix of channels to maximize the total response given a set of response functions and constraints.}
\usage{
mrm_optimal_mix(
  mrm,
  total,
  x0 = NULL,
  lb = NULL,
  ub = NULL,
  ineq_constr = NULL
)
}
\arguments{
\item{mrm}{A list of response functions, where each function takes a numeric vector as input and returns a numeric value.}

\item{total}{A numeric value representing the total budget or constraint for the optimization.}

\item{x0}{An optional numeric vector representing the initial guess for the optimization. If NULL, a default value will be used.}

\item{lb}{An optional numeric vector representing the lower bounds for each channel. If NULL, default values will be used.}

\item{ub}{An optional numeric vector representing the upper bounds for each channel. If NULL, default values will be used.}

\item{ineq_constr}{An optional function representing additional inequality constraints. If NULL, default constraints will be used.}
}
\value{
A list containing the optimization results, including the optimal channel mix and the maximum response value.
}
\description{
Optimal Mix Function
This function computes the optimal mix of channels to maximize the total response given a set of response functions and constraints.
}
