% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_returns_ranges.R
\name{mrm_returns_ranges}
\alias{mrm_returns_ranges}
\title{Get Ranges of Maximum Marginal Returns and Average Returns
This function computes the ranges of x values corresponding to the maximum marginal returns (MR)
and average returns (AR) for a list of marginal response models (MRMs).}
\usage{
mrm_returns_ranges(mrm, xrange = NULL)
}
\arguments{
\item{mrm}{A list of fitted model objects (e.g., brmsfit objects).}

\item{xrange}{A numeric vector of length 2 specifying the range of x values to consider. If NULL, the range is determined from the data.}
}
\value{
A data frame containing the channels and their corresponding x ranges for maximum MR and AR.
}
\description{
Get Ranges of Maximum Marginal Returns and Average Returns
This function computes the ranges of x values corresponding to the maximum marginal returns (MR)
and average returns (AR) for a list of marginal response models (MRMs).
}
\details{
The function computes the MR and AR for each model in the list and identifies the x values where MR and AR are maximized. It returns a data frame with the channel names and their respective x ranges.
}
