% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_range.R
\name{plot_range}
\alias{plot_range}
\title{Plot the response rate curves}
\usage{
plot_range(rc_fit, avg_cp = 1)
}
\arguments{
\item{rc_fit}{A response curve fit object.}

\item{avg_cp}{The average cost per response. Default is 1.}
}
\value{
A ggplot object representing the response rate curves.
}
\description{
This function plots the response rate curves for a given response curve fit.
}
