% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrm_tv_snapshot.R
\name{print.mrmfit_tv_snapshot}
\alias{print.mrmfit_tv_snapshot}
\title{Print method for mrmfit_tv_snapshot objects}
\usage{
\method{print}{mrmfit_tv_snapshot}(x, ...)
}
\arguments{
\item{x}{A \code{mrmfit_tv_snapshot} object from \code{\link{mrm_tv_snapshot}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for mrmfit_tv_snapshot objects
}
