% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_variables_in_formula.R
\name{replace_variables_in_formula}
\alias{replace_variables_in_formula}
\title{Replace Variables in a Formula}
\usage{
replace_variables_in_formula(formula, old_vars, new_vars)
}
\arguments{
\item{formula}{A formula object where variables need to be replaced.}

\item{old_vars}{A character vector of variable names to be replaced.}

\item{new_vars}{A character vector of new variable names to replace the old ones.}
}
\value{
A formula object with the specified variables replaced.
}
\description{
This function replaces variables in a given formula with new variable names.
}
\details{
The function uses regular expressions to match the variable names in the formula.
}
