% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Function to compute standard error from model}
\usage{
se(f, data)
}
\arguments{
\item{f}{Formula for model}

\item{data}{Data.frame of data}
}
\value{
Standard error value
}
\description{
Function to compute standard error from model
}
\examples{
d <- mrc::simdat(n = 100, r = 0.4)
se(y ~ x1 + x2, data = d)

}
