% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dafault_constraints.R
\name{set_dafault_constraints}
\alias{set_dafault_constraints}
\title{Set default constraints for optimization}
\usage{
set_dafault_constraints(C, total)
}
\arguments{
\item{C}{Number of variables}

\item{total}{Total sum constraint}
}
\value{
A list containing initial values (x0), lower bounds (lb), upper bounds (ub), and a total constraint function (total_constr_func).
}
\description{
This function sets default constraints for an optimization problem, including initial values, lower and upper bounds, and a total constraint function.
}
