% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_total_constraint.R
\name{set_total_constraint}
\alias{set_total_constraint}
\title{Set a total constraint function}
\usage{
set_total_constraint(total)
}
\arguments{
\item{total}{A numeric value representing the desired total sum of the elements in the vector.}
}
\value{
A function that takes a numeric vector as input and returns a numeric vector of constraints.
}
\description{
This function creates a constraint function that ensures the sum of the elements in a vector equals a specified total.
}
