% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-simulation.R
\name{simcovmat}
\alias{simcovmat}
\title{Simulate covariance matrix}
\usage{
simcovmat(n, r)
}
\arguments{
\item{n}{Number of data points for raw data}

\item{r}{Rho. Correlation value to be input into 2x2 positive-definite correlation matrix.}
}
\value{
Covariance matrix
}
\description{
Simulate covariance matrix
}
\examples{
simcovmat(n = 100, r = 0.4)

}
