% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulate correlated data}
\usage{
simdat(n, r)
}
\arguments{
\item{n}{Number of datapoints}

\item{r}{Rho. Correlation value to be input into 2x2 positive-definite correlation matrix}
}
\value{
Returns data.frame with 'y' dependent variable and 'x1' and 'x2' as independent variables
}
\description{
Simulate correlated data
}
