% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr.R
\name{spr}
\alias{spr}
\title{Calculate semi-partial correlations from data}
\usage{
spr(f, vint, data)
}
\arguments{
\item{f}{Formula for full model}

\item{vint}{Independent variable of interest}

\item{data}{Data frame correlation matrix}
}
\value{
Semi-partial correlation value
}
\description{
Calculate semi-partial correlations from data
}
