% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr.R
\name{sprf}
\alias{sprf}
\title{Calculate semi-partial correlations from data by specifying a formula.}
\usage{
sprf(f, data)
}
\arguments{
\item{f}{Formula for entire model in which label\link{1} is your independent variable of interest.}

\item{data}{Data.frame of data}
}
\value{
Semi-partial correlation value
}
\description{
Calculate semi-partial correlations from data by specifying a formula.
}
