% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{tsr}
\alias{tsr}
\title{Function that calculates t-value for semi-partial correlation}
\usage{
tsr(f, data)
}
\arguments{
\item{f}{Formula for entire model in which label\link{1} is your independent variable of interest.}

\item{data}{Data.frame of data}
}
\value{
T-test with t, semi-partial correlation and standard error
}
\description{
Function that calculates t-value for semi-partial correlation
}
\examples{
d <- mrc::simdat(n = 100, r = 0.4)
tsr(y ~ x1 + x2, data = d)

}
