% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr.R
\name{spr}
\alias{spr}
\title{Calculate semi-partial correlations from data}
\usage{
spr(dep, ivi, ivn)
}
\arguments{
\item{dep}{Dependent variable}

\item{ivi}{Independent variable of interest}

\item{ivn}{Independent variable of NO interest}
}
\value{
Semi-partial correlation value
}
\description{
Calculate semi-partial correlations from data
}
\examples{
d <- mrc::simdata(n = 100, r = 0.4)
spr(dep = d$y, ivi = d$x1, ivn = d$x2)

}
